+++ CHANGELOG.md
@@ -4,7 +4,11 @@ entry.
 
 ## 13.0.13 (2020-08-17)
 
-- No changes.
+### Security (2 changes)
+
+- Stop deploy token being mis-used as user in ProjectPolicy and GroupPolicy.
+- Project access is checked during deploy token authentication.
+
 
 ## 13.0.12 (2020-08-05)